Initial attempt at building DS9 with nix
This seems to work - in that I get a ds9 executable that runs - but
I have not checked the build log for possible errors (some were
seen during development and I don't know if I have all of them),
and I have not put the resulting binary through its paces.

I will note that the fonts look ugly.